# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Freelance Marketplace API
  description: A comprehensive API for managing a freelance marketplace platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@freelancehub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://freelancehub.com/api
    description: Production server
security:
  - BearerAuth: []
  - SessionAuth: []
paths:
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: "john@example.com"
      responses:
        '200':
          description: Password reset link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /projects/{id}:
    get:
      tags:
        - Projects
      summary: Get project by ID
      description: Retrieve a specific project with full details
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Projects
      summary: Update project
      description: Update a project (requires ownership or admin role)
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                images:
                  type: array
                  items:
                    type: string
                status:
                  type: string
                  enum: [ACTIVE, INACTIVE, COMPLETED]
                featured:
                  type: boolean
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Projects
      summary: Delete project
      description: Delete a project (requires ownership or admin role)
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieve a specific order
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Orders
      summary: Update order
      description: Update order status (requires seller or admin role)
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PENDING, PAID, COMPLETED, CANCELLED]
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [USER, ADMIN]
          example: "USER"
        bio:
          type: string
          example: "Experienced full-stack developer"
        skills:
          type: array
          items:
            type: string
          example: ["React", "Node.js", "TypeScript"]
        rating:
          type: number
          minimum: 0
          maximum: 5
          example: 4.8
        totalReviews:
          type: integer
          minimum: 0
          example: 15
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: string
          example: "proj_123"
        title:
          type: string
          example: "E-commerce Website Development"
        description:
          type: string
          example: "Build a complete e-commerce website with modern technologies"
        price:
          type: number
          minimum: 0
          example: 2500
        category:
          type: string
          example: "Web Development"
        tags:
          type: array
          items:
            type: string
          example: ["React", "Node.js", "PostgreSQL"]
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/image1.jpg"]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, COMPLETED]
          example: "ACTIVE"
        featured:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        author:
          $ref: '#/components/schemas/User'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        _count:
          type: object
          properties:
            orders:
              type: integer
            reviews:
              type: integer
    Order:
      type: object
      properties:
        id:
          type: string
          example: "order_123"
        status:
          type: string
          enum: [PENDING, PAID, COMPLETED, CANCELLED]
          example: "PAID"
        totalAmount:
          type: number
          minimum: 0
          example: 2500
        paymentId:
          type: string
          example: "pi_123"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        buyer:
          $ref: '#/components/schemas/User'
        project:
          $ref: '#/components/schemas/Project'
        review:
          $ref: '#/components/schemas/Review'
    Review:
      type: object
      properties:
        id:
          type: string
          example: "review_123"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Excellent work! Highly recommended."
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        reviewer:
          $ref: '#/components/schemas/User'
        project:
          $ref: '#/components/schemas/Project'
        order:
          $ref: '#/components/schemas/Order'
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token